Move all ownership related routes to krate::owner::*
